requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
python-decouple>=3.8
Pillow>=10.0.0
